Support building on Qt6 while keeping Qt5 (dual build)
Make apm_planner build and link against both Qt5 and Qt6 so the project
can transition off end-of-life Qt5 gradually. Verified: both
`cmake -DCMAKE_PREFIX_PATH=$(brew --prefix qt)` (Qt6) and `qt@5` (Qt5)
configure, build, and produce a working apmplanner2.app.

Build system:
- Root, opmapcontrol, and quazip CMakeLists use the version-agnostic
  `find_package(QT NAMES Qt6 Qt5)` + `Qt${QT_VERSION_MAJOR}` pattern.
- Bump C++ standard 14 -> 17 (required by Qt6, valid for Qt5).
- Qt6 also links Core5Compat (legacy QRegExp/QTextCodec), OpenGLWidgets
  (OpenGL split in Qt6), and SvgWidgets (QGraphicsSvgItem moved there).

Source ports (all dual-safe across Qt5 and Qt6):
- QString methods taking QRegExp (split/replace/contains/indexOf) are
  removed in Qt6 and NOT bridged by Core5Compat -> ported those call
  sites to QRegularExpression; configuration.h now returns
  QRegularExpression. Pure QRegExp-class files just include <QRegExp>.
- toTime_t/fromTime_t/setTime_t -> *SecsSinceEpoch.
- QString::SkipEmptyParts -> Qt::SkipEmptyParts.
- QPainter::HighQualityAntialiasing -> Antialiasing.
- QLayout::setMargin -> setContentsMargins; QFileDialog::DirectoryOnly ->
  Directory; QSortFilterProxyModel::setFilterRegExp ->
  setFilterRegularExpression; qSort -> std::sort; trUtf8 -> tr.
- QTextStream::setCodec guarded -> setEncoding(QStringConverter::Utf8);
  QTextCodec dropped (QString::fromUtf8) / linked via Core5Compat.
- QGLWidget viewport -> QOpenGLWidget + QSurfaceFormat; drop dead
  QGLWidget/QDesktopWidget includes; add explicit QStandardPaths include.
- Explicit QChar() for int->QChar and bool/QString concatenations now
  that QChar(int) is explicit in Qt6; cast qsizetype narrowing.
- Guard `using namespace QtDataVisualization` (namespace gone in Qt6).
- QGC.h: replace isnan/isinf macros (clobbered QtQml's std::isinf) with
  `using std::isnan/isinf`.
- qcustomplot.h: skip the Q_GADGET-in-namespace QCP trick under the
  Qt 6.9+ moc (its enum reflection is unused here).

+++ src/ui/Loghandling/LogdataStorage.cpp
@@ -566,7 +566,7 @@ QStringList LogdataStorage::setupUnitData(const QString &timeStampName, double d
                         // find the max index within the first 50 entries and store it within the datatype
                         const auto &valueRow {m_dataStorage[type.m_name]};
                         int maxIndex {0};
-                        int maxEntriesToCheck {valueRow.size() < s_maxItemsToCheck ? valueRow.size() : s_maxItemsToCheck};
+                        int maxEntriesToCheck {static_cast<int>(valueRow.size() < s_maxItemsToCheck ? valueRow.size() : s_maxItemsToCheck)};
 
                         for (int i = 0; i < maxEntriesToCheck; ++i)
                         {
